+++ scrapy_poet/downloadermiddlewares.py
@@ -2,6 +2,7 @@
 responsible for injecting Page Input dependencies before the request callbacks
 are executed.
 """
+
 import inspect
 import logging
 import warnings
@@ -13,6 +14,7 @@ from scrapy.downloadermiddlewares.stats import DownloaderStats
 from scrapy.http import Request, Response
 from twisted.internet.defer import Deferred, inlineCallbacks
 from web_poet import RulesRegistry
+from web_poet.exceptions import Retry
 
 from .api import DummyResponse
 from .injection import Injector
@@ -154,10 +156,25 @@ class InjectionMiddleware:
             return response
 
         # Find out the dependencies
-        final_kwargs = yield from self.injector.build_callback_dependencies(
-            request,
-            response,
-        )
+        try:
+            final_kwargs = yield from self.injector.build_callback_dependencies(
+                request,
+                response,
+            )
+        except Retry as exception:
+            # Needed for Twisted < 21.2.0. See the discussion thread linked below:
+            # https://github.com/scrapinghub/scrapy-poet/pull/129#discussion_r1102693967
+            from scrapy.downloadermiddlewares.retry import get_retry_request
+
+            reason = str(exception) or "page_object_retry"
+            new_request_or_none = get_retry_request(
+                request,
+                spider=spider,
+                reason=reason,
+            )
+            if not new_request_or_none:
+                return response
+            return new_request_or_none
         # Fill the callback arguments with the created instances
         for arg, value in final_kwargs.items():
             # If scrapy-poet can't provided the dependency, allow the user to